+++ packages/react-cosmos-plugin-vite/src/indexHtml/findMainScriptUrl.test.ts
