+++ packages/react-cosmos-plugin-vite/src/indexHtml/generateViteIndexHtml.ts
